Introduce a new protocol AnyLayout to reduce repeating code in this package and also in the vapor-provider package.
